moved makeRequest into its own module
